provide dedicated routines to handle codePoint{At,After,Before}(char[])
the java being replaced here works -- `new String(char[])` produces a string of those chars;
but it transpiles in JS to `new String(string[])` which does not do the same thing,
as JS uses the `toString` on the array and inserts a comma between all the characters.

(transpiling `new String(char[] a)` should trigger `new String(a?.join(''))` in TS;
but that is a separate issue.)

this PR simply copies the code from string to chars with minor tweaks for code-point and length/end.

i'm not a unicode expert however so i'm not 100% sure of it, but it works for me in simple cases --
whereas the old code only ever worked for strings of length <=1! (this is also skips the full-string-copy
which will be much faster.)